Use tls/x509 validation
Currently, in both the server and the client we validate the other side's
certificates by comparing them directly with each of the allowed ones.

While this works, it is a bit tedious, somewhat risky, and it means we don't
support delegation via signing.

This patch changes the code to use Go's libraries for validating the
certificates, which makes the code simpler (specially the client), and also
paves the road for more features.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>