tests: Include an OpenSSL configuration for our CA operations
When doing CA operations, OpenSSL relies significantly on the system
configuration in a way that make our tests depend on it (for example, with the
directory where the CA database is kept).

To make the tests more independent of the local OpenSSL configuration, this
patch includes its own, based on Debian's default.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>