tests: Use $LOGNAME instead of os.getlogin()
The tests use os.getlogin() to get the user they're running under, which is
used just for informational names to make troubleshooting easier.

However, that function fails when the controlling terminal isn't a tty; this
is triggered for example by the Debian automated build process, which is how
this bug was originally reported (http://bugs.debian.org/755036).

To fix that, we can just try to get $LOGNAME from the environment, and fall
back to "unknown" if missing, which is harmless.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>