tests: Assert negotiated cipher secret size >= 128 bits
Currently, the tests assert that the negotiated cipher secret size is > 128
bits. Due to recent changes (probably in openssl), in Debian testing we end up
negotiating ECDHE-RSA-AES128-GCM-SHA256 which has a 128 bit key.

That causes the test to fail.

There's nothing wrong with 128 bit keys in this context, and probably the
original intention was to use >= but got stuck with > as it worked.

While the value of this check is arguable, let's keep it for now, just
changing the check to >= so 128 bit secrets don't cause the test to fail.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>