tests: Fix SSL validation in test_tricky
test_tricky uses httplib to create a client, which used to not validate the
server certificate.

Python 2.7.9 changes that, and now the test fail because the client cannot
validate the server.

The problem is that to fix this, we need to use the new "context" parameter
which is not backwards-compatible. So we have to add a little version-specific
code to work around this.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>