tests: Use shorter organizationalUnitName
The organizationalUnitName is currently created based on usernames and
hostnames, but those can get too long for openssl, causing unnecessary tests
errors, and breaking hermicity.

This patch fixes this by making organizationalUnitName use shorter names,
which are just as useful as the variable ones in practice.

Note we still need them to be different for server, client and CA, otherwise
openssl complains in some cases.