tests: Comment out test that changed behaviour during go 1.8
Go commit 8ad70a549364c216c386afcead4dccfdcd39748b by Adam Langley
(agl@golang.org) changed the behaviour of how leaf certificates on root work
on Go.

Before, they would not be assumed valid and we had a test for it.
But the change makes the library consider them explicitly valid, so our test
breaks.

However, this is not something we expect to be a problem for our users, in
practice it's a really unusual corner case for kxd.

The change is not even in a released Go version yet, so changing the tests
would cause them to fail on released versions. But leaving as-is will cause
them to fail when go 1.8 is released.

So we leave the test commented out for now, we can re-enable it once there are
a couple of Go versions with the new behaviour.