kxd: Implement a hook for running an external command on key requests
This patch makes kxd (optionally) run a hook on each key request.

The hook is an external script (its location is given via a command line
flag). If it exits with code 0, then the request is allowed. Otherwise,
it is rejected.

It is run right before sending email notifications, after the request
has been approved.

Information about the request (key, certificate chain validation, remote
address, etc.) is passed via the environment.

This can be used to implement more elaborate access checks and
notifications.