Require Go >= 1.11
Debian stable (which we generally use as reference) was recently
released with Go 1.11, and bumping the version allows us to enable a few
tests that were not valid in <= 1.7.

So bump the required version to 1.11, which is higher than strictly
needed as per current code, but by now it should be a reasonable
requirement.