Use our own tool to generate certificates
This patch introduces an kxgencert tool that can be used to generate
self-signed certificates.

All scripts and tests are adjusted to use it, removing the dependency on
openssl for certificate manipulation.

Unfortunately, the CA tests are removed, although the functionality
continues to be transparently supported.

The main reason is that the way we were generating the self-signed
certificates is not compatible with Go >= 1.15, since the usage of the
CommonName field is deprecated:
https://golang.org/doc/go1.15#commonname.

Because of that, also most clients built with Go 1.15 will no longer
talk to servers with old certificates. This can be worked around by
setting the GODEBUG=x509ignoreCN=0 environment variable.