cryptsetup: Copy /etc/kxc into initramfs only if it exists
Today, the initramfs hook that copies /etc/kxc does so unconditionally.

That fails if the directory doesn't exist, for example if kxc was
installed but not configured yet.

This patch fixes the issue by only doing the copy if the directory
exists.

Thanks to Leandro Lucarella for reporting this bug:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1063700