Minimum supported Go version: 1.21
Update the minimum supported Go version to 1.21 (which was released 1y
ago), and update the GitHub Actions config accordingly.

While we're not currently using any of the specific features right now,
the current minimum version of Go 1.15 is a bit misleading since it's
quite old and not really being tested against in practice.