tests: Add tests for DNSNames='*'
In kxd <= 0.16, by default kxgencert (and before than, the equivalent
scripts) would generate certificates with DNSNames='*', and everything
worked okay.

But in Go 1.23, the Go TLS library started to reject such certificates
(Go commit 375031d8dcec9ae74d2dbc437b201107dba3bb5f).

We still want to make sure that the server can handle such certificates,
as they might still be in use in the wild, so this patch adds an
explicit test for that scenario.

Note that the tests or the code hasn't been updated to fix the problem
yet, but having this test will ensure that kxd can handle these certs
after the fixes.