kxd: Use standard representation for cert name
When logging certificate subject (for human reference), today we only
print the organizational unit. That may be too brief for some certs, and
make troubleshooting more difficult.

This patch changes the logic to use the RFC 2253 standard representation
for RDN sequences, which is much more comprehensive and common, while
still being reasonably compact for logging purposes.